if문 중첩 변경
if문을 2번만 중첩되게 변경함
